chatbot-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.1
openai==1.53.0
requests==2.32.3
